requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
pyyaml>=6.0

